requirements.txt
# Core Trading Libraries
ccxt>=4.2.25                    # Unified exchange API
google-generativeai>=0.8.3      # Gemini API client (Google AI)

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Data Analysis
pandas>=2.2.0
numpy>=1.26.0,<2.0.0            # NumPy 2.0 may have compatibility issues
ta>=0.11.0                      # Technical analysis library

# Backtesting
vectorbt>=0.26.0                # Vectorized backtesting

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
streamlit>=1.31.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0                   # Better logging
aiohttp>=3.9.0                  # Async HTTP client
websockets>=12.0                # WebSocket client
tenacity>=8.2.0                 # Retry logic
pytz>=2024.1                    # Timezone handling

# Monitoring & Notifications
python-telegram-bot>=20.8       # Telegram notifications

# Development (more flexible versions)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0

# Optional: Advanced Features
# ta-lib==0.4.28                # Requires separate C library installation
# quantstats>=0.0.62            # Portfolio analytics
# anthropic>=0.18.1             # Claude API client (optional alternative to Gemini)
